make loading simpler + make it optimal for model indexing
I think this repo that we are building to focus more in inference
can double up as model indexer that we can later use in NBX. Simply
load the model and deploy it anywhere with 3 lines.

This package can act as a mega-centralised repository with all the
models along with all the highly optimised kernels in other languages
as well.